Split Reflection and SourceGen TypeInfos
