+++ src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/JsonMetadataServicesConverter.cs
@@ -50,24 +50,7 @@ namespace System.Text.Json.Serialization.Converters
         }
 
         internal override bool OnTryRead(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options, ref ReadStack state, out T? value)
-        {
-            JsonTypeInfo jsonTypeInfo = state.Current.JsonTypeInfo;
-
-            if (_converterStrategy == ConverterStrategy.Object)
-            {
-                if (jsonTypeInfo.PropertyCache == null)
-                {
-                    jsonTypeInfo.InitializePropCache();
-                }
-
-                if (jsonTypeInfo.ParameterCache == null && jsonTypeInfo.IsObjectWithParameterizedCtor)
-                {
-                    jsonTypeInfo.InitializeParameterCache();
-                }
-            }
-
-            return Converter.OnTryRead(ref reader, typeToConvert, options, ref state, out value);
-        }
+            => Converter.OnTryRead(ref reader, typeToConvert, options, ref state, out value);
 
         internal override bool OnTryWrite(Utf8JsonWriter writer, T value, JsonSerializerOptions options, ref WriteStack state)
         {
@@ -82,11 +65,6 @@ namespace System.Text.Json.Serialization.Converters
             {
                 info.SerializeHandler(writer, value);
                 return true;
-            }
-
-            if (_converterStrategy == ConverterStrategy.Object && jsonTypeInfo.PropertyCache == null)
-            {
-                jsonTypeInfo.InitializePropCache();
             }
 
             return Converter.OnTryWrite(writer, value, options, ref state);
