+++ src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonSerializer.Write.Stream.cs
@@ -256,6 +256,7 @@ namespace System.Text.Json
             using (var writer = new Utf8JsonWriter(bufferWriter, writerOptions))
             {
                 WriteStack state = new WriteStack { CancellationToken = cancellationToken };
+                jsonTypeInfo.EnsureConfigured();
                 JsonConverter converter = state.Initialize(jsonTypeInfo, supportContinuation: true, supportAsync: true);
 
                 bool isFinalBlock;
@@ -329,6 +330,7 @@ namespace System.Text.Json
             using (var writer = new Utf8JsonWriter(bufferWriter, writerOptions))
             {
                 WriteStack state = default;
+                jsonTypeInfo.EnsureConfigured();
                 JsonConverter converter = state.Initialize(jsonTypeInfo, supportContinuation: true, supportAsync: false);
 
                 bool isFinalBlock;
