+++ src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Metadata/JsonMetadataServices.Collections.cs
@@ -19,7 +19,7 @@ namespace System.Text.Json.Serialization.Metadata
         /// <returns>Serialization metadata for the given type.</returns>
         /// <remarks>This API is for use by the output of the System.Text.Json source generator and should not be called directly.</remarks>
         public static JsonTypeInfo<TElement[]> CreateArrayInfo<TElement>(JsonSerializerOptions options, JsonCollectionInfoValues<TElement[]> collectionInfo)
-            => new JsonTypeInfoInternal<TElement[]>(
+            => new SourceGenJsonTypeInfo<TElement[]>(
                 options,
                 collectionInfo,
                 () => new ArrayConverter<TElement[], TElement>());
@@ -37,7 +37,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonSerializerOptions options,
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : List<TElement>
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new ListOfTConverter<TCollection, TElement>());
@@ -57,7 +57,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : Dictionary<TKey, TValue>
             where TKey : notnull
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new DictionaryOfTKeyTValueConverter<TCollection, TKey, TValue>());
@@ -83,7 +83,7 @@ namespace System.Text.Json.Serialization.Metadata
             Func<IEnumerable<KeyValuePair<TKey, TValue>>, TCollection> createRangeFunc!!)
             where TCollection : IReadOnlyDictionary<TKey, TValue>
             where TKey : notnull
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new ImmutableDictionaryOfTKeyTValueConverter<TCollection, TKey, TValue>(),
@@ -104,7 +104,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : IDictionary<TKey, TValue>
             where TKey : notnull
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new IDictionaryOfTKeyTValueConverter<TCollection, TKey, TValue>());
@@ -124,7 +124,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : IReadOnlyDictionary<TKey, TValue>
             where TKey : notnull
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new IReadOnlyDictionaryOfTKeyTValueConverter<TCollection, TKey, TValue>());
@@ -146,7 +146,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonCollectionInfoValues<TCollection> collectionInfo,
             Func<IEnumerable<TElement>, TCollection> createRangeFunc!!)
             where TCollection : IEnumerable<TElement>
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new ImmutableEnumerableOfTConverter<TCollection, TElement>(),
@@ -164,7 +164,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonSerializerOptions options,
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : IList
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new IListConverter<TCollection>());
@@ -182,7 +182,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonSerializerOptions options,
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : IList<TElement>
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new IListOfTConverter<TCollection, TElement>());
@@ -200,7 +200,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonSerializerOptions options,
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : ISet<TElement>
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new ISetOfTConverter<TCollection, TElement>());
@@ -218,7 +218,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonSerializerOptions options,
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : ICollection<TElement>
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new ICollectionOfTConverter<TCollection, TElement>());
@@ -236,7 +236,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonSerializerOptions options,
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : Stack<TElement>
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new StackOfTConverter<TCollection, TElement>());
@@ -254,7 +254,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonSerializerOptions options,
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : Queue<TElement>
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new QueueOfTConverter<TCollection, TElement>());
@@ -272,7 +272,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonSerializerOptions options,
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : ConcurrentStack<TElement>
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new ConcurrentStackOfTConverter<TCollection, TElement>());
@@ -290,7 +290,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonSerializerOptions options,
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : ConcurrentQueue<TElement>
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new ConcurrentQueueOfTConverter<TCollection, TElement>());
@@ -308,7 +308,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonSerializerOptions options,
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : IEnumerable<TElement>
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new IEnumerableOfTConverter<TCollection, TElement>());
@@ -325,7 +325,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonSerializerOptions options,
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : IDictionary
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new IDictionaryConverter<TCollection>());
@@ -371,7 +371,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonCollectionInfoValues<TCollection> collectionInfo,
             Action<TCollection, object?> addFunc!!)
             where TCollection : IEnumerable
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new StackOrQueueConverter<TCollection>(),
@@ -390,7 +390,7 @@ namespace System.Text.Json.Serialization.Metadata
             JsonSerializerOptions options,
             JsonCollectionInfoValues<TCollection> collectionInfo)
             where TCollection : IEnumerable
-            => new JsonTypeInfoInternal<TCollection>(
+            => new SourceGenJsonTypeInfo<TCollection>(
                 options,
                 collectionInfo,
                 () => new IEnumerableConverter<TCollection>());
