+++ src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Metadata/JsonTypeInfo.Cache.cs
@@ -51,37 +51,6 @@ namespace System.Text.Json.Serialization.Metadata
 
         internal Func<JsonParameterInfoValues[]>? CtorParamInitFunc;
 
-        internal static JsonPropertyInfo AddProperty(
-            MemberInfo memberInfo,
-            Type memberType,
-            Type parentClassType,
-            bool isVirtual,
-            JsonNumberHandling? parentTypeNumberHandling,
-            JsonSerializerOptions options)
-        {
-            JsonIgnoreCondition? ignoreCondition = JsonPropertyInfo.GetAttribute<JsonIgnoreAttribute>(memberInfo)?.Condition;
-            if (ignoreCondition == JsonIgnoreCondition.Always)
-            {
-                return JsonPropertyInfo.CreateIgnoredPropertyPlaceholder(memberInfo, memberType, isVirtual, options);
-            }
-
-            JsonConverter converter = GetConverter(
-                memberType,
-                parentClassType,
-                memberInfo,
-                options);
-
-            return CreateProperty(
-                declaredPropertyType: memberType,
-                memberInfo,
-                parentClassType,
-                isVirtual,
-                converter,
-                options,
-                parentTypeNumberHandling,
-                ignoreCondition);
-        }
-
         internal static JsonPropertyInfo CreateProperty(
             Type declaredPropertyType,
             MemberInfo? memberInfo,
@@ -113,7 +82,7 @@ namespace System.Text.Json.Serialization.Metadata
         /// Create a <see cref="JsonPropertyInfo"/> for a given Type.
         /// See <seealso cref="PropertyInfoForTypeInfo"/>.
         /// </summary>
-        internal static JsonPropertyInfo CreatePropertyInfoForTypeInfo(
+        private static JsonPropertyInfo CreatePropertyInfoForTypeInfo(
             Type declaredPropertyType,
             JsonConverter converter,
             JsonNumberHandling? numberHandling,
@@ -142,6 +111,7 @@ namespace System.Text.Json.Serialization.Metadata
         {
             PropertyRef propertyRef;
 
+            ValidateCanBeUsedForDeserialization();
             ulong key = GetKey(propertyName);
 
             // Keep a local copy of the cache in case it changes by another thread.
@@ -568,93 +538,6 @@ namespace System.Text.Json.Serialization.Metadata
             }
 
             frame.CtorArgumentState.ParameterRefCache = null;
-        }
-
-        internal void InitializePropCache()
-        {
-            Debug.Assert(PropertyInfoForTypeInfo.ConverterStrategy == ConverterStrategy.Object);
-
-            // Delayed JsonTypeInfo initialization can be invoked by
-            // multiple threads, add a temporary check to avoid contention.
-            // TODO refactor so that metadata initialization is single threaded.
-            if (Volatile.Read(ref PropertyCache) is not null)
-            {
-                return;
-            }
-
-            JsonSerializerContext? context = Options.JsonSerializerContext;
-            Debug.Assert(context != null);
-
-            JsonPropertyInfo[] array;
-            if (PropInitFunc == null || (array = PropInitFunc(context)) == null)
-            {
-                ThrowHelper.ThrowInvalidOperationException_NoMetadataForTypeProperties(context, Type);
-                return;
-            }
-
-            Dictionary<string, JsonPropertyInfo>? ignoredMembers = null;
-            JsonPropertyDictionary<JsonPropertyInfo> propertyCache = new(Options.PropertyNameCaseInsensitive, array.Length);
-
-            for (int i = 0; i < array.Length; i++)
-            {
-                JsonPropertyInfo jsonPropertyInfo = array[i];
-                bool hasJsonInclude = jsonPropertyInfo.SrcGen_HasJsonInclude;
-
-                if (!jsonPropertyInfo.SrcGen_IsPublic)
-                {
-                    if (hasJsonInclude)
-                    {
-                        ThrowHelper.ThrowInvalidOperationException_JsonIncludeOnNonPublicInvalid(jsonPropertyInfo.ClrName!, jsonPropertyInfo.DeclaringType);
-                    }
-
-                    continue;
-                }
-
-                if (jsonPropertyInfo.MemberType == MemberTypes.Field && !hasJsonInclude && !Options.IncludeFields)
-                {
-                    continue;
-                }
-
-                if (jsonPropertyInfo.SrcGen_IsExtensionData)
-                {
-                    Debug.Assert(IsValidDataExtensionProperty(jsonPropertyInfo));
-
-                    DataExtensionProperty = jsonPropertyInfo;
-                    continue;
-                }
-
-                CacheMember(jsonPropertyInfo, propertyCache, ref ignoredMembers);
-            }
-
-            // Populate a local cache and assign it to the global cache after completion.
-            Volatile.Write(ref PropertyCache, propertyCache);
-        }
-
-        internal void InitializeParameterCache()
-        {
-            Debug.Assert(PropertyCache != null);
-            Debug.Assert(PropertyInfoForTypeInfo.ConverterStrategy == ConverterStrategy.Object);
-
-            // Delayed JsonTypeInfo initialization can be invoked by
-            // multiple threads, add a temporary check to avoid contention.
-            // TODO refactor so that metadata initialization is single threaded.
-            if (Volatile.Read(ref ParameterCache) is not null)
-            {
-                return;
-            }
-
-            JsonSerializerContext? context = Options.JsonSerializerContext;
-            Debug.Assert(context != null);
-
-            JsonParameterInfoValues[] array;
-            if (CtorParamInitFunc == null || (array = CtorParamInitFunc()) == null)
-            {
-                ThrowHelper.ThrowInvalidOperationException_NoMetadataForTypeCtorParams(context, Type);
-                return;
-            }
-
-            InitializeConstructorParameters(array, sourceGenMode: true);
-            Debug.Assert(ParameterCache != null);
         }
     }
 }
